Home Button: Add fake home button
OS-312, OS-330
Introduce a fake home button to close the app. Ideally this would
utilise the actual home button but the mechanism to combine the two
doesn't work so just add a button in the app. As this closes the app, it
prevents the home button from returning to the dashboard in the
background.